Updated gitIgnore.

+++ src/main/java/com/samcox/ranker/media/AddMediaRequest.java
@@ -1,14 +1,14 @@
 package com.samcox.ranker.media;
 
 public class AddMediaRequest {
-  private int tmdbId;
+  private Long tmdbId;
   private MediaType mediaType;
   private int ranking;
 
   public Long getTmdbId() {
     return tmdbId;
   }
-  public void setTmdbId(int tmdbId, MediaType mediaType, int ranking) {
+  public void setTmdbId(Long tmdbId, MediaType mediaType, int ranking) {
     this.tmdbId = tmdbId;
     this.mediaType = mediaType;
     this.ranking = ranking;
